+++ tests/custom_termination.rs
@@ -1,9 +1,10 @@
 use chrome_for_testing_manager::*;
+use rootcause::Report;
 
 mod common;
 
 #[tokio::test(flavor = "multi_thread")]
-async fn custom_termination() -> anyhow::Result<()> {
+async fn custom_termination() -> Result<(), Report> {
     tracing_subscriber::fmt().try_init().ok();
 
     // NOTE: Using beta channel as stable channel chromedriver was bugged on Linux...