+++ tests/single_session_non_headless.rs
@@ -1,10 +1,11 @@
 use chrome_for_testing_manager::*;
+use rootcause::Report;
 use thirtyfour::prelude::*;
 
 mod common;
 
 #[tokio::test(flavor = "multi_thread")]
-async fn single_session_non_headless() -> anyhow::Result<()> {
+async fn single_session_non_headless() -> Result<(), Report> {
     tracing_subscriber::fmt().try_init().ok();
 
     // NOTE: Using beta channel as stable channel chromedriver was bugged on Linux...
@@ -14,5 +15,7 @@ async fn single_session_non_headless() -> anyhow::Result<()> {
             |caps| caps.unset_headless(),
             common::wikipedia::test_wikipedia,
         )
-        .await
+        .await?;
+
+    Ok(())
 }